Fixed the java<->utils circular import
