+++ tests/uhs_unittest.py
@@ -23,8 +23,8 @@ import tempfile
 import unittest
 
 from openquake import java
+from openquake.java import list_to_jdouble_array
 from openquake.shapes import Site
-from openquake.utils import list_to_jdouble_array
 from openquake.hazard.uhs.core import (
     compute_uhs, touch_result_file, write_uhs_results)
